10/15

+++ WebContent/ServiceCenter/Q&Aboard/content.jsp
@@ -12,6 +12,8 @@
 </head>
 <body>
 <%
+   String id =(String)session.getAttribute("id");
+ out.print(id);
    int num = Integer.parseInt(request.getParameter("num")); 
    String pageNum = request.getParameter("pageNum"); 
    //웹헤더에서 넘어온 num값과 pagenum값  선언과 생성
@@ -60,24 +62,30 @@
   </tr>
   <tr height="30">      
     <td colspan="4" align="right" > 
-   <input type="button" value="글수정" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_updateForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
-       <!-- num ,pagenum값을 가지고 updateform.jsp로 넘어간다 -->
-    &nbsp;&nbsp;&nbsp;&nbsp;
-   <input type="button" value="글삭제" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_deleteForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
-     <!-- num ,pagenum값을 가지고 deleteForm.jsp로 넘어간다 -->
-    &nbsp;&nbsp;&nbsp;&nbsp;
-      <input type="button" value="답글쓰기" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_writeForm.jsp?num=<%=num%>&ref=<%=ref%>&re_step=<%=re_step%>&re_level=<%=re_level%>'">
-        <!-- num ,ref,re_step,re_level을  가지고 writeForm.jsp로 넘어간다 -->
-    &nbsp;&nbsp;&nbsp;&nbsp;
-       <input type="button" value="글목록" 
-       onclick="document.location.href='/ServiceCenter/Q&Aboard/Q&A.jsp?pageNum=<%=pageNum%>'">
-  <!--pagenum을  가지고 list.jsp로 넘어간다 -->
+    <%if("admin".equals(session.getAttribute("id"))){ %>
+    <%if(session.getAttribute("id").equals(article.getWriter())){%>  
+	   <input type="button" value="글수정" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_updateForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+	 	  &nbsp;&nbsp;&nbsp;&nbsp;
+	   <input type="button" value="글삭제" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_deleteForm.jsp?num=<%=article.getNum()%>&pageNum=<%=pageNum%>'">
+			&nbsp;&nbsp;&nbsp;&nbsp;
+			<input type="button" value="글목록" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/Q&A.jsp?pageNum=<%=pageNum%>'">
+	    	&nbsp;&nbsp;&nbsp;&nbsp; 
+		<%} %>
+		
+	      <input type="button" value="답글쓰기" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/qna_writeForm.jsp?num=<%=num%>&ref=<%=ref%>&re_step=<%=re_step%>&re_level=<%=re_level%>'">
+	      &nbsp;&nbsp;&nbsp;&nbsp;
+		<%} %>
+	       <input type="button" value="글목록" 
+	       onclick="document.location.href='/ServiceCenter/Q&Aboard/Q&A.jsp?pageNum=<%=pageNum%>'">
+	    	&nbsp;&nbsp;&nbsp;&nbsp; 
     </td>
   </tr>
-</table>    
+</table>  
+
 <%
  }catch(Exception e){} 
  %>
